reorder app folder
