+++ src/types/auth/service.type.d.ts
