+++ src/types/overview/enum.type.ts
@@ -1,0 +1,4 @@
+export enum DiscountTypes {
+  DIRECT = "direct",
+  ACCUMULATIVE = "accumulative",
+}
