+++ src/types/overview/service.type.d.ts
@@ -1,4 +1,5 @@
-import { OverviewModel } from "./model.type";
+import type { OverviewModel } from "./model.type";
+
 export interface OverviewApiService {
   getOverview(): Promise<OverviewModel.Sale>;
 }
